Add files via upload
This code features examples showing how to create a matrix with all possible combinations of variables with varying values. It also shows how to multiply each row of the matrix by a vector. This would be useful for tasks such as creating different weightings for a each variable in an index. This could would allow you to create all combinations of indices and multiply them by the raw data to see which produces the best results.